requirements.txt
annotated-types==0.5.0
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.99.1
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pydantic==1.10.10
pydantic_core==2.0.1
pymongo==3.11.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarg==0.1.9
redis~=4.6.0